requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
croniter==6.0.0
discord-py-interactions==5.14.0
discord-typings==0.9.0
discord.py==2.4.0
emoji==2.14.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.1.0
numpy==2.2.2
propcache==0.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
